added menu item view functionality
I added the review model for access to the db. Also, I make it so when
you look at a menu item, if your a retail account you can make comments
but only as the retail user and not your actual user account so people
know its the restaurant making the comments. Lastly, I made the
transition so if you click on a menu item it will take you to the proper
page